factor out getSavedClineMessages
